requirements.txt
python-dotenv==0.21.0
numpy==1.24.2
pandas==1.5.3
jupyter==1.0.0
yfinance==0.2.12
beautifulsoup4==4.11.1
pyarrow==11.0.0
scikit-learn==1.2.1